Render custom poster and toggle back to main
In this commit, I added a handler function to handle
the logic of rendering the user generated poster.
This is accomplished by making use of the makeCustomPoster
function to update the currentPoster variable. The form
and main poster section visibility is then toggled.
The currentPoster is rendered by passing in the
corresponding key values to the poster rendering functions.

Finally, I added a click event listener to the
"Show my poster" button that invokes the above
mentioned handler function.