Remove posters from saved posters
In this commit the removePosterFromArray function accesses the event
target's id number, compares it against every element's unique id, and
if it finds a match, splices it out of the savedPosters array.

After this function in the double click event listener attached
to the saved posters grid, the renderSavedPosters function is invoked
to re-render the saved posters grid with the new array of saved posters.